Removes LastCommittedTxIdSetter
LastCommittedTxIdSetter was a hook used by HA and Zookeeper to update the latest tx id for each instance on commit, to be used in master election. That is no longer required in the new HA code.
Introduces a generic LastTxIdGetter to expose similar, pull based functionality for retrieving the latest tx id of a database.